+++ NEWS.md
@@ -1,3 +1,5 @@
+* fix bug found by DB that prevents `evaluate_at()` to work if
+  no strata are provided to a bag (#10)
 * fix bug that prevents `evaluate_at()` to work if data supplied 
   strata not in the eqs_bag.